Add handling for all links
Since this will be in a walled garden, the other outbound links on the
page (e.g. PushPay) won't work either unless we auth the user first.

This adds handling for all links on the page, authing them and sending
them to the href of the link.

We can then leverage this machinery for the continue button as well,
just by settings its href to the user_continue_url Meraki gives us.