Provide a fallback URL for when continue_url is missing
I'm not exactly sure how this is happening but visitors are sometimes
being sent to apple.com when they click continue. I'm guessing these
might be cases where the continue URL is missing? I'm not sure what the
continue URL would be set to when an iPhone hits the apple
hotspot-detect site, but I could see it being blank in that case.

In any case, this should provide a reasonable next page any time the JS
doesn't find a continue_url with which to replace the link target.